+++ phpBB2/install/schemas/mysql_basic.sql
@@ -4,6 +4,8 @@
 # $Id: mysql_basic.sql,v 1.29.2.25 2006/03/09 21:55:09 grahamje Exp $
 
 # -- Config
+INSERT INTO phpbb_user_id_sequence (singleton, last_id) VALUES (1, 0);
+
 INSERT INTO phpbb_config (config_name, config_value) VALUES ('config_id','1');
 INSERT INTO phpbb_config (config_name, config_value) VALUES ('board_disable','0');
 INSERT INTO phpbb_config (config_name, config_value) VALUES ('sitename','yourdomain.com');